+++ src/class01/Code02_SelectionSort.java
@@ -1,2 +1,4 @@
-package class01;public class Code02_SelectionSort {
+package class01;
+
+public class Code02_SelectionSort {
 }
